refactoring to output namespace
